Add new Platform* header files. Scripts run on mac now.
Just an initial attempt at some platform specific stuff. For now,
it is only for getting the scripts dir. Current code depended on
the current_dir+"scripts/", but current_dir might not be correct,
like in the case of mac app bundles.

This just adds a function GetScriptsPath(), with platform specific
implementations. For Linux, it should probably be updated to use
getcwd().

Also, todo: the linux and windows implementations should really
return an absolute path, and then rename this method to indicate that.

Feel free to refactor to a better approach to handle this. pimpl-like?